Trying to fix viewport clicking to aim.

+++ src/com/stewsters/weapons/Weapon.java
@@ -6,9 +6,9 @@ import com.stewsters.Person;
 public interface Weapon {
 
 
-    public void leftClick(Person holder, int xClick, int yClick);
+    public void leftClick(Person holder, float xClick, float yClick);
 
-    public void rightClick(Person holder, int xClick, int yClick);
+    public void rightClick(Person holder, float xClick, float yClick);
 
     public void leftRelease();
 }
